chore : 불필요 코드 제거 및 소스코드 및 클래스명 변경
- 리뉴얼 작업으로 더 이상 사용하지 않은 contentdetailViewModel 리소스 삭제
