--- lib/domain/model/content/content_shell.dart
@@ -1,45 +1,0 @@
-import 'package:soon_sak/utilities/index.dart';
-
-/** Created By Ximya - 2022.12.31
- *  컨텐츠 간략 정보를 담고 있는 데이터 모델
- *   [DTO]
- *   {
-    "id": 113988,
-    "posterImgUrl": "/f2PVrphK0u81ES256lw3oAZuF3x.jpg",
-    "type": "tv"
-    }
- * */
-
-class ContentPosterShell {
-  final String originId;
-  final int contentId;
-  final ContentType contentType;
-  final String posterImgUrl;
-  final String title;
-
-  ContentPosterShell({
-    required this.title,
-    required this.originId,
-    required this.contentId,
-    required this.contentType,
-    required this.posterImgUrl,
-  });
-
-  factory ContentPosterShell.fromResponse(TopTenContentItemResponse response) =>
-      ContentPosterShell(
-        title: response.title,
-        originId: response.id,
-        contentId: SplittedIdAndType.fromOriginId(response.id).id,
-        contentType: SplittedIdAndType.fromOriginId(response.id).type,
-        posterImgUrl: response.imgUrl,
-      );
-
-// // Form Mock Json Data Response
-// factory ContentPosterShell.fromJson(Map<String, dynamic> json) {
-//   return ContentPosterShell(
-//     contentId:  SplittedIdAndType.fromOriginId(json['id']).id,
-//     posterImgUrl: json['posterImgUrl'],
-//     contentType: SplittedIdAndType.fromOriginId(json['id']).type,
-//     );
-// }
-}
