+++ lib/presentation/screens/contentDetail/controllerResources/content_detail_info_tab_view_model.part.dart
@@ -2,34 +2,12 @@ part of '../content_detail_view_model.dart';
 
 /** Crated By Ximya - 2022.12.10
  * 컨텐츠 상세 페이지 > 정보 탭뷰 영역 컨트롤러 리소스
+ *
+ * Edited By Ximya - 2023.06.09
+ * 리뉴얼 작어브로 불필요 리소스 삭제
  * */
 
 extension ContentDetailInfoTabViewModel on ContentDetailViewModel {
-  // 컨텐츠 이미지 리스트
-  List<String>? get contentImgList => contentImgUrlList;
-
-  // 컨텐츠 이미지 존재 여부 (일부 컨텐츠에서 컨텐츠 이미지가 존재 하지 않음)
-  bool? get contentImgExist {
-    return contentImgUrlList?.isNotEmpty ?? true ? true : false;
-  }
-
-  // // 총 조회수
-  // String? get totalViewCount => Formatter.formatNumberWithUnit(
-  //       videoInfo?.mainViewCount,
-  //       isViewCount: false,
-  //     );
-  //
-  // // 좋아요 수
-  // String? get likesCount {
-  //   return Formatter.formatNumberWithUnit(videoInfo?.mainLikesCount);
-  // }
-
-  // 유튜브 컨텐츠 업로드 일자
-  String? get youtubeUploadDate =>
-      Formatter.getDateDifferenceFromNow(videoInfo?.mainUploadDate.valueOrNull);
-
-  // 채널 설명
-  String? get channelDescription => channelInfo?.description;
 
   // 채널 구독자 수
   int? get subscriberCount =>
@@ -39,21 +17,9 @@ extension ContentDetailInfoTabViewModel on ContentDetailViewModel {
   String? get channelImgUrl =>
       _passedArgument.channelLogoImgUrl ?? channelInfo?.logoImgUrl;
 
-  // 채널 명
-  String? get channelName => _passedArgument.channelName ?? channelInfo?.name;
-
-  /// 방영일
-  /// 컨텐츠 타입이 영화일 경우 방영상태 대신 방영일을 보여줌
-  String? get contentReleaseDate => _contentDescriptionInfo?.releaseDate;
-
-  // 방영상태
-  String? get contentAirStatus => _contentDescriptionInfo?.airStatus;
-
   // 출연진 정보 개수
   int? get creditCount => contentCreditList?.length;
 
-  // 큐레이터 정보
-  UserModel? get curator => _curator;
 
   // 출연진 섹션 > [CarouselSlider] 개수
   int? get sliderCount {
