--- lib/presentation/screens/contentDetail/localWidget/tabView/content_detail_info_sacffold.dart
@@ -1,10 +1,0 @@
-import 'package:soon_sak/utilities/index.dart';
-
-class ContentDetailInfoScaffold extends StatelessWidget {
-  const ContentDetailInfoScaffold({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
