+++ lib/presentation/screens/profileSetting/profile_setting_binding.dart
@@ -1,8 +1,8 @@
-import 'package:soon_sak/app/di/custom_binding.dart';
+import 'package:soon_sak/app/di/binding.dart';
 import 'package:soon_sak/presentation/screens/profileSetting/profile_setting_view_model.dart';
 import 'package:soon_sak/utilities/index.dart';
 
-class ProfileSettingBinding extends CustomBindings {
+class ProfileSettingBinding extends Bindings {
   @override
   void dependencies() {
     super.dependencies();
