+++ lib/presentation/screens/search/search_binding.dart
@@ -1,9 +1,9 @@
-import 'package:soon_sak/app/di/custom_binding.dart';
+import 'package:soon_sak/app/di/binding.dart';
 import 'package:soon_sak/domain/useCase/search/new_search_paged_content_use_case.dart';
 import 'package:soon_sak/presentation/screens/search/localWidget/search_scaffold_controller.dart';
 import 'package:soon_sak/utilities/index.dart';
 
-class SearchBinding extends CustomBindings {
+class SearchBinding extends Bindings {
   @override
   void dependencies() {
     super.dependencies();
