+++ lib/presentation/screens/setting/setting_binding.dart
@@ -1,8 +1,8 @@
-import 'package:soon_sak/app/di/custom_binding.dart';
+import 'package:soon_sak/app/di/binding.dart';
 import 'package:soon_sak/presentation/screens/setting/setting_view_model.dart';
 import 'package:soon_sak/utilities/index.dart';
 
-class SettingBinding extends CustomBindings {
+class SettingBinding extends Bindings {
   @override
   void dependencies() {
     super.dependencies();
