+++ pubspec.yaml
@@ -4,7 +4,7 @@ description: A new Flutter project.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.4+9
+version: 1.0.5+10
 
 environment:
   sdk: '>=2.18.2 <3.0.0'
@@ -33,7 +33,7 @@ dependencies:
   flutter_staggered_grid_view: ^0.6.2 # 그리드뷰
   visibility_detector: ^0.4.0+2 #위젯 노출 여부 감지
   dots_indicator: ^2.1.0 #dot 인디케이터
-  adaptive_dialog: ^1.8.3+1 # 모달
+
 
 
   # State Controller
